Write function showing winner of 5 rounds between player and cpu
The function game() calls the playRound() function 5 times using loop.
Two variables keep score tally of player and cpu.
Each round, the player is prompted to choose between rock paper and
scissors. The result of each round and cumul score is logged in the
console.
When the five rounds are finished, the final winner is logged in the
console.